Added base 'quarantine' module which sets up the module path and whatnot.
This way we can point people to add this line to their .bashrc:

    module load /path/to/quarantine/modules/quarantine

In that module I've defined and exported the environment variable QUARANTINE_PATH, which
now the other modules depend on to locate the quarantine. In this way, the quarantine
can be relocated and the path needs only be set up in one place.